Removed debugging and band-aid fixed server bind
Removed printf debugging in server.c, also fixed an issue where if a server exited with clients connected via SIGINT. The server couldn't bind to that port until the clients pressed "Enter" key to refresh.

For now, the server just binds to the port above if a bind fails. If the second bind fails, an error is thrown. Clients who cannot connect to a server now try the port above too.